scripts/I&W_Document_Generator/requirements.txt
pandas
requests
urllib3
python-docx
pytz
threatconnect
RequestObject